Health and Bank UI components

See merge request course-work/cgra252/2023/assignments/taliposabi/assignment-1-unreal!2

+++ Source/LittleMario/Public/PickUp/LMCoinPickUp.h
@@ -14,9 +14,12 @@ class LITTLEMARIO_API ALMCoinPickUp : public ALMBasePickUp
 {
 	GENERATED_BODY()
 
+	//public:
+ //     float GetCoinsAmount() const { return CoinsAmount; }
+
 	protected:
       UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "PickUp", meta = (ClampMin = "1.0", ClampMax = "10.0"))
-      int32 CoinsAmount = 10;
+      float CoinsAmount = 10;
 
 	private:
       virtual bool GivePickupTo(APawn* PlayerPawn) override;
